% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_parameter_combinations.R
\name{generate_parameter_combinations}
\alias{generate_parameter_combinations}
\title{generate_parameter_combinations}
\usage{
generate_parameter_combinations(default_parameters, parameters_to_tweak)
}
\arguments{
\item{default_parameters}{list of the default parameters}

\item{parameters_to_tweak}{the parameter to tweak}
}
\value{
a matrix containing the possible combinations of the parameters
}
\description{
generate_parameter_combinations
}
