% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_points.R
\name{generate_points}
\alias{generate_points}
\title{Generate new points from density matrix}
\usage{
generate_points(density, num_points = 1000)
}
\arguments{
\item{density}{A matrix with density values.}

\item{num_points}{The number of points to sample}
}
\description{
Generate new points from density matrix
}
