% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GNG.R
\name{GNG}
\alias{GNG}
\title{GNG}
\usage{
GNG(
  x,
  max_iter,
  epsilon_b,
  epsilon_n,
  age_max,
  max_nodes,
  lambda,
  alpha,
  beta,
  verbose = TRUE,
  make_logs_at = NULL
)
}
\arguments{
\item{x}{the expression dataset}

\item{max_iter}{Maximum of iterations}

\item{epsilon_n}{The percentage by which the neighbors of the closest node will move}

\item{age_max}{Maximum age of the edges}

\item{max_nodes}{Maximum of nodes in the graph}

\item{lambda}{iteration where new nodes can be inserted}

\item{alpha}{error of the closest node}

\item{beta}{percentage by which the errors of all nodes will decrease at every iteration}

\item{verbose}{boolean}

\item{make_logs_at}{At which iterations to store the GNG, for visualisation purposes.}

\item{The}{percentage by which the closest node will move}
}
\value{
A gng object
}
\description{
The Growing Neural Gas algorithm adapted to trajectory inference
}
