% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_trajectory_svg.R
\name{read_trajectory_svg}
\alias{read_trajectory_svg}
\title{Read trajectory svg}
\usage{
read_trajectory_svg(svg_file, pdf_file, png_file)
}
\arguments{
\item{id}{The name of the trajectory svg}
}
\description{
Read trajectory svg
}
