% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_methods.R
\name{run_methods}
\alias{run_methods}
\title{run_methods}
\usage{
run_methods(dataset, method_param_name, parameters)
}
\arguments{
\item{dataset}{The scRNAseq dataset, as returned by dynwrap}

\item{method_param_name}{Name of the method to be used (ex = gng_param())}

\item{parameters}{The parameters that will be given to the method param}
}
\value{
a list of errors and results from the required GNG version
}
\description{
run_methods
}
