% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualise_datasets.R
\name{visualise_datasets}
\alias{visualise_datasets}
\title{visualise datasets}
\usage{
visualise_datasets(files, directory)
}
\arguments{
\item{files}{the files to be visualised, as passed to the dynbenchmark::load_dataset() function}

\item{directory}{the directory where the pngs should be saved}
}
\value{
The plots for all files
}
\description{
visualise datasets
}
\examples{
visualise_datasets(files = c("synthetic/dyntoy/cyclic_1","synthetic/dyntoy/bifurcating_4"),
                   directory = "plots/")
}
